move new methods in rdvinsertion api endpoints
