+++ spec/requests/api/rdvinsertion/creneau_availability_request_spec.rb
@@ -1,6 +1,6 @@
 require "swagger_helper"
 
-describe "Available Creneaux Count for Invitation" do
+describe "Available Creneaux Count for Invitation", swagger_doc: "v1/api.json" do
   with_examples
 
   path "/api/rdvinsertion/invitations/creneau_availability" do
